pyproject.toml
[tool.poetry]
name = "hoolieats"
version = "0.1.0"
description = ""
authors = ["gamberooni <zhaoyoong@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.9.7 || >3.9.7,<4.0"
dbt-duckdb = "^1.2.0"
duckdb = "^0.5.1"
jupyter = "^1.0.0"
pandas = "^1.5.0"
numpy = "^1.23.3"
pre-commit = "^2.20.0"
streamlit = "^1.13.0"
plotly-express = "^0.4.1"
watchdog = "^2.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
